TST: clean up temporarily installed packages
