+++ tests/test_install.py
@@ -1,33 +1,8 @@
-import pkg_resources
-import subprocess
-import sys
-
 import pytest
 
 import audobject
 
-
-def uninstall(
-    package: str,
-    module: str,
-):
-    # uninstall package
-    subprocess.check_call(
-        [
-            sys.executable,
-            '-m',
-            'pip',
-            'uninstall',
-            '--yes',
-            package,
-        ]
-    )
-    # remove module
-    for m in list(sys.modules):
-        if m.startswith(package):
-            sys.modules.pop(m)
-    # force pkg_resources to re-scan site packages
-    pkg_resources._initialize_master_working_set()
+from conftest import uninstall
 
 
 @pytest.mark.parametrize(
@@ -62,6 +37,7 @@ def uninstall(
     ],
 )
 def test(yaml_s, package, module):
+    # fails because of missing packages
     with pytest.raises(ModuleNotFoundError):
         audobject.from_yaml_s(
             yaml_s,
